Re-enabled achievements (#415)
* Re-enabled achievements

Fixes that PV was triggered by launching a rocket. Why on earth was PV internally just a new locale slapped on the the vanilla rocket launch achievement?

This also enables other achievements that were incidentally disabled by a temp fix a year and a half ago. Some of them aren't completely finished and might need polish.

* Update changelog.txt